+++ CHANGELOG.md
@@ -92,6 +92,7 @@ Version 1.5.0
 - fix alpha draw order issues in Hendrix, Generator_Small01, Generator_Small02, LavaTester01, Weapon_RiotShield meshes (RF bug)
 - fix corona lights being visible through clutters, items, corpses (RF bug)
 - fix possible 'buzzing' sound in multi-player game if some players are in water (RF bug)
+- fix dedicated server crash when loading level that uses directional light (RF bug)
 
 Version 1.4.1 (released 2019-12-07)
 -----------------------------------